requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.1.5
django-bootstrap3==14.2.0
django-crispy-forms==1.10.0
django-debug-toolbar==3.2
django-taggit==1.3.0
django-widget-tweaks==1.4.8
isort==5.7.0
Pillow==8.1.0
python-dotenv==0.15.0
pytz==2020.5
sorl-thumbnail==12.7.0
soupsieve==2.1
sqlparse==0.4.1
